Thicken raster ticks. Move highlight span drawing to axes layer (below main content) from cursor layer.
