+++ src/views/RasterPlot/TimeScrollView/TSVAxesLayer.tsx
@@ -1,12 +1,13 @@
 import BaseCanvas from 'FigurlCanvas/BaseCanvas';
 import React, { useMemo } from 'react';
 import { paintAxes } from './paint';
-import { TimeScrollViewPanel, TimeTick } from './TimeScrollView';
+import { PixelHighlightSpanSet, TimeScrollViewPanel, TimeTick } from './TimeScrollView';
 
 export type TSVAxesLayerProps<T extends {[key: string]: any}> = {
     panels: TimeScrollViewPanel<T>[]
     timeRange: [number, number]
     timeTicks: TimeTick[]
+    highlightSpans?: PixelHighlightSpanSet[]
     margins: {left: number, right: number, top: number, bottom: number}
     selectedPanelKeys: string[]
     panelHeight: number
@@ -17,10 +18,10 @@ export type TSVAxesLayerProps<T extends {[key: string]: any}> = {
 }
 
 const TSVAxesLayer = <T extends {[key: string]: any}>(props: TSVAxesLayerProps<T>) => {
-    const {width, height, panels, panelHeight, perPanelOffset, timeRange, timeTicks, margins, selectedPanelKeys, hideTimeAxis} = props
+    const {width, height, panels, panelHeight, perPanelOffset, timeRange, timeTicks, highlightSpans, margins, selectedPanelKeys, hideTimeAxis} = props
     const drawData = useMemo(() => ({
-        width, height, panels, panelHeight, perPanelOffset, timeRange, timeTicks, margins, selectedPanelKeys, hideTimeAxis
-    }), [width, height, panels, panelHeight, perPanelOffset, timeRange, timeTicks, margins, selectedPanelKeys, hideTimeAxis])
+        width, height, panels, panelHeight, perPanelOffset, timeRange, timeTicks, highlightSpans, margins, selectedPanelKeys, hideTimeAxis
+    }), [width, height, panels, panelHeight, perPanelOffset, timeRange, timeTicks, highlightSpans, margins, selectedPanelKeys, hideTimeAxis])
 
     return (
         <BaseCanvas
